requirements.txt
streamlit
Pillow
streamlit-lottie
requests